+++ tests/unit/util/editor/interactions/creationSnapping.spec.ts
@@ -1,9 +1,14 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 import { LineString, Polygon } from 'ol/geom.js';
-import { Math as CesiumMath, Cartesian2 } from '@vcmap-cesium/engine';
+import {
+  Math as CesiumMath,
+  Cartesian2,
+  HeightReference,
+} from '@vcmap-cesium/engine';
 import {
   CesiumMap,
+  createVertex,
   EventType,
   ModificationKeyType,
   OpenlayersMap,
@@ -296,6 +301,53 @@ describe('creation snapping', () => {
         positionOrPixel: [1, 2.1, 12],
       });
       arrayCloseTo(modifiedEvent.positionOrPixel!, [1.05, 2.05, 12]);
+    });
+
+    describe('with altitude mode absolute', () => {
+      before(() => {
+        scratchLayer.vectorProperties.altitudeMode = HeightReference.NONE;
+      });
+
+      after(() => {
+        scratchLayer.vectorProperties.altitudeMode =
+          HeightReference.CLAMP_TO_GROUND;
+      });
+
+      it('should not snap on top of itself, if absolute', async () => {
+        const geometry = new LineString([
+          [0, 0, 0],
+          [1, 1, 0],
+          [0, 1, 0],
+          [1, 1.5, 0],
+        ]);
+        creationSnapping.setGeometry(geometry);
+        const modifiedEvent = await creationSnapping.pipe({
+          ...eventBase,
+          map: cesiumMap,
+          type: EventType.DRAG,
+          position: [0.5, 0.5, 12],
+          positionOrPixel: [0.5, 0.5, 12],
+        });
+        arrayCloseTo(modifiedEvent.positionOrPixel!, [0.5, 0.5, 12]);
+      });
+
+      it('should snap on top of itself, if absolute but 2D coordinates', async () => {
+        const geometry = new LineString([
+          [0, 0, 0],
+          [1, 1, 0],
+          [0, 1, 0],
+          [1, 1.5, 0],
+        ]);
+        creationSnapping.setGeometry(geometry);
+        const modifiedEvent = await creationSnapping.pipe({
+          ...eventBase,
+          map: cesiumMap,
+          type: EventType.DRAG,
+          position: [0.5, 0.6],
+          positionOrPixel: [0.5, 0.6],
+        });
+        arrayCloseTo(modifiedEvent.positionOrPixel!, [0.55, 0.55]);
+      });
     });
   });